[Kernel] Vectorize fp8 quant layernorm and fix input_stride type
- Use q8_n_t<fp8_type, VEC_SIZE> for vectorised FP8 output writes in
  rms_norm_static_fp8_quant_kernel (previously scalar per-element).
- Use vectorize_read_with_alignment<VEC_SIZE> for variance pass.
- Fix input_stride from int64_t to int in fp8 quant kernels and host
  launchers (stride fits in 32 bits).
- Use local contiguous_input handle in rms_norm_static_fp8_quant to
  avoid mutating the caller's tensor reference.

+++ csrc/layernorm_kernels.cu
@@ -218,7 +218,7 @@ void rms_norm(torch::Tensor& out,     // [..., hidden_size]
   VLLM_DISPATCH_RANK234(num_dims, [&] {
     VLLM_DISPATCH_FLOATING_TYPES(input.scalar_type(), "rms_norm_kernel", [&] {
       const int calculated_vec_size =
-          std::gcd(16 / sizeof(scalar_t), hidden_size);
+          std::gcd(static_cast<int>(16 / sizeof(scalar_t)), hidden_size);
       const int block_size =
           std::min(hidden_size / calculated_vec_size, max_block_size);
       dim3 block(block_size);
